GCP/requirements.txt
aniso8601==9.0.1
click==8.1.5
Flask==2.2.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
gunicorn==20.0.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.6
zipp==3.16.2
google-cloud-datastore==2.15.1